chore(deps): remove unused Babel dependencies
Removed @babel/core, @babel/preset-env, and babel-jest as they are not
used in the project. Jest is configured to use ts-jest for TypeScript
transformation, making these Babel dependencies unnecessary.

Note: The deprecated glob@7.2.3 and inflight warnings still appear
because ts-jest itself depends on @jest/transform which pulls in
babel-plugin-istanbul. This is a transitive dependency issue that will
be resolved when the Jest ecosystem updates to newer glob versions.